Fixed issue related to newer MATLAB versions
This addresses and error occurring for MATLAB R2018a (encountered on Ubuntu 64-bit) "error: cannot convert ‘int*’.... " perhaps this can be fixed in a more intelligent way (e.g. see statements on using mwSize in thread below) but I fixed it using the 32-bit compatibility flag -DMX_COMPAT_32 as suggested in this thread https://stackoverflow.com/questions/43045701/switch-compiler-on-matlab-2017a-in-ubuntu-os. 
I also fixed a minor typo "this may TAKE some time" and altered the file so it can be run independent of the current working directory.